Un-embed Config field from internal services
As it might be confusing if these services export the Config fields as
well.

The link struct still embeds the Config as it makes it easier to access'
the configuration and it will used through the interface instead of the
link struct's signature and thus hopefully won't lead to confusion.